Update models.py
I believe this is the correct implementation for having a collection of Models in another Model. There is no need to have the collection of Models as a field of another Model. So we don't need a list of Player_Years as a field a Player, we just need we just need the Player Foreign Key Field in each Player_Year. The link to the doc can explain this better.

https://docs.djangoproject.com/en/1.4/topics/db/examples/many_to_one/